requirements.txt
Django==1.7.1
Markdown==2.5.1
django-admin-bootstrapped==2.1.0
django-filter==0.8
djangorestframework==2.4.3
wsgiref==0.1.2
